moved ip editor to separate file
