+++ I2I/MainWindow.h
@@ -3,8 +3,6 @@
 
 #include <QMainWindow>
 #include <QTcpServer>
-#include <QFrame>
-#include <QLineEdit>
 #include "User.h"
 #include "ChatController.h"
 
@@ -30,40 +28,6 @@ private:
     QList<ChatController*> chats;
 };
 
-
-//=============================================================================
-class CustomIpEditor : public QFrame
-{
-    Q_OBJECT
-public:
-    explicit CustomIpEditor(QWidget *parent = 0);
-
-    virtual ~CustomIpEditor() {}
-};
-
-//=============================================================================
-class CustomLineEdit : public QLineEdit
-{
-    Q_OBJECT
-public:
-    explicit CustomLineEdit(const QString & contents = "", QWidget *parent = 0);
-    virtual ~CustomLineEdit() {}
-
-signals:
-    void jumpForward();
-    void jumpBackward();
-
-public slots:
-    void jumpIn();
-
-protected:
-    virtual void focusInEvent(QFocusEvent *event);
-    virtual void keyPressEvent(QKeyEvent * event);
-    virtual void mouseReleaseEvent(QMouseEvent *event);
-
-private:
-    bool selectOnMouseRelease;
-};
 
 
 #endif // MAINWINDOW_H
